requirements_file.txt
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.0.74
Flask==2.3.2
Pillow==9.5.0
numpy==1.24.3
Werkzeug==2.3.6
